TNB_HACKATHON/TNB/requirements.txt
braintree==3.38.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.6
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
idna==2.6
jsonpickle==0.9.5
mysqlclient==1.3.12
oauthlib==2.0.6
PyJWT==1.5.3
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
six==1.11.0
django-allauth==0.34.0
aiohttp==2.3.7
pyfcm==1.4.3
django-countries
django-gulp
Pillow

